requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
backoff==2.2.1
    # via huawei-solar
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
huawei-solar==2.5.0
    # via -r requirements.in
idna==3.11
    # via requests
influxdb==5.3.2
    # via -r requirements.in
influxdb_client==1.49.0
    # via -r requirements.in
msgpack==1.1.2
    # via influxdb
pymodbus==3.11.4
    # via huawei-solar
pyserial==3.5
    # via pyserial-asyncio
pyserial-asyncio==0.6
    # via huawei-solar
python-dateutil==2.9.0.post0
    # via influxdb
pytz==2025.2
    # via
    #   huawei-solar
    #   influxdb
pyyaml==6.0.3
    # via -r requirements.in
requests==2.32.5
    # via influxdb
six==1.17.0
    # via
    #   influxdb
    #   python-dateutil
urllib3==2.6.0
    # via requests
